Add metadata ingestion
A mostly complete first draft of creating the metadata for the WRI power plant data.

First incomplete item: we don't have a good S3 bucket location for the metadata store.  Please advise (and fix).

Second incomplete item: writing the metadata to the store should contemplate the possibility that there's already tables with metadata present, in which case this new metadata either needs to be inserted or existing metadata updated.  The logic of what to do will become clear once we have existing metadata to play with.